WIP improve SkillTabs and add other ideas
